+++ api/src/test/java/org/cardanofoundation/rosetta/api/construction/preprocess/PreprocessApiTest.java
@@ -39,7 +39,7 @@ public class PreprocessApiTest extends IntegrationTest {
 
     ConstructionPreprocessResponse constructionPreprocessResponse = constructionApiService.constructionPreprocessService(
         preprocessRequest);
-    Map<String, Double> options = (Map<String, Double>)constructionPreprocessResponse.getOptions();
+    Map<String, Integer> options = (Map<String, Integer>)constructionPreprocessResponse.getOptions();
     assertEquals(1000, options.get("relative_ttl"));
     assertEquals(224, options.get("transaction_size"));
   }
@@ -52,7 +52,7 @@ public class PreprocessApiTest extends IntegrationTest {
 
     ConstructionPreprocessResponse constructionPreprocessResponse = constructionApiService.constructionPreprocessService(
         preprocessRequest);
-    Map<String, Double> options = (Map<String, Double>)constructionPreprocessResponse.getOptions();
+    Map<String, Integer> options = (Map<String, Integer>)constructionPreprocessResponse.getOptions();
     assertEquals(100, options.get("relative_ttl"));
     assertEquals(399, options.get("transaction_size"));
   }
@@ -65,7 +65,7 @@ public class PreprocessApiTest extends IntegrationTest {
 
     ConstructionPreprocessResponse constructionPreprocessResponse = constructionApiService.constructionPreprocessService(
         preprocessRequest);
-    Map<String, Double> options = (Map<String, Double>)constructionPreprocessResponse.getOptions();
+    Map<String, Integer> options = (Map<String, Integer>)constructionPreprocessResponse.getOptions();
     assertEquals(100, options.get("relative_ttl"));
     assertEquals(921, options.get("transaction_size"));
   }
